requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
python-multipart==0.0.7
pydantic==2.5.3
python-dotenv==1.0.0
jose==0.1.0